+++ index.js
@@ -2,6 +2,7 @@
  * @format
  */
 
+import 'react-native-url-polyfill/auto';
 import { AppRegistry } from 'react-native';
 import notifee from '@notifee/react-native';
 import App from './App';